+++ packages/synapse-interface/components/layouts/LandingPageWrapper/index.tsx
@@ -54,7 +54,7 @@ export function LandingPageWrapper({ children }: { children: any }) {
           bannerId="2024-05-02-scroll-launch"
           bannerContents="Synapse Protocol now available on Scroll"
           startDate={new Date('2024-05-08T18:45:09+00:00')}
-          endDate={new Date('2024-05-27T18:45:09+00:00')}
+          endDate={new Date('2024-06-15T18:45:09+00:00')}
         />
         <MaintenanceBanners />
         <LandingNav />